requirements.txt
fastapi
uvicorn
pydantic<3
python-multipart
psycopg2-binary
